Adjusts sentence extaction alg for less whitespace
Letters like i and j are not a continuous stroke of ink, and so we must
continue to parse whitespace when we come across the first dot by
checking if there's "ink" in the next consecutive pixels. However, doing
so prematurely begins collecting rows of empty whitespace before
actually getting to the letter. Therefore we only do this if we've
already collected a nonwhite pixel row.